requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0
chromadb>=0.4.0
groq>=0.4.0
python-pptx>=0.6.21
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
